+++ client/src/interface/board.ts
